requirements.txt
beautifulsoup4==4.11.2
cloudscraper==1.2.68
discord.py==2.1.1
pycoingecko==3.1.0
